Use -Wshorten-64-to-32 if it's available, and fix warnings it shows.
Probably harmless, but do it to help catch some issues on LP64 (64-bit
UN*X) or LLP64 (64-bit Windows) platforms when compiling on macOS.

+++ rpcapd/fileconf.c
@@ -202,7 +202,7 @@ static int strrem(char *string, char chr)
 {
 	char *pos;
 	int num = 0;
-	int len, i;
+	size_t len, i;
 
 	while ((pos = strchr(string, chr)) != NULL)
 	{
